distill-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
groq==0.9.0
pydub==0.25.1
python-multipart==0.0.9
pydantic-settings==2.4.0
python-dotenv==1.0.1
